Ability to pause on next, or specified, layer change.
LC_PAUSE will toggle on/off the pause on the next layer flag. If no LAYER is passed, the flag will be set to pause on the next layer change. If a LAYER is passed, a pause on layer change will occur when the slicer's passed layer value (_LAYER_CHANGE LAYER={layer_num}) is equal to the user's LAYER passed with this macro.